Split `doc.rs` into modules
